move market table to a reusable component
